[CCAP-897] Adjusting some things for adding multiple providers (#1569)

+++ src/main/java/org/ilgcc/app/submission/conditions/HasProviderForOneChild.java
@@ -2,7 +2,6 @@ package org.ilgcc.app.submission.conditions;
 
 import static org.ilgcc.app.utils.SubmissionUtilities.getChildrenNeedingAssistance;
 import static org.ilgcc.app.utils.SubmissionUtilities.hasChosenProvider;
-import static org.ilgcc.app.utils.SubmissionUtilities.hasNotChosenProvider;
 
 import formflow.library.config.submission.Condition;
 import formflow.library.data.Submission;
@@ -18,7 +17,7 @@ public class HasProviderForOneChild implements Condition {
     }
     @Override
     public Boolean run(Submission submission) {
-        boolean oneChildNeedsChildcareAssistance = (long) getChildrenNeedingAssistance(submission).size() == 1;
+        boolean oneChildNeedsChildcareAssistance = getChildrenNeedingAssistance(submission).size() == 1;
         return enableMultipleProviders && hasChosenProvider(submission) && oneChildNeedsChildcareAssistance;
     }
 }
